Refactor: Update ReaderSettings to use 'Courier New' font and enhance settings modal with custom CSS and JS tabs
